Refactor response parsing and scanner (#167)
Previously there were two different ResultSets, one for rowwise data
and another for the newer columnar format. This has been collapsed into
a single ResultSet that supports the columnar format. A subsequent PR
will add support for transforming the rowwise format into the columnar
one, to support older versions of Kudu.

RowResult has been changed from a `ref struct` to a `struct`. VB.NET
doesn't support ref structs, and this also simplifies consuming scan
results. Additionally, ResultSet now implements `IEnumerable<RowResult>`.

Parsing results from Kudu has been improved. Parsing has been moved from
KuduConnection to a dedicated Protocol namespace. CPU usage has been
improved for scanning by using System.IO.Pipeline's new ReadAtLeastAsync()
to reduce the number of continuations.

The default scan batch size has been increased from 1mb to 8mb. This
provides a nice performance boost by reducing the number of round trips
to the Kudu server. It also pairs nicely with the recently uncapped
ArrayPool in .NET 6.

+++ test/Knet.Kudu.Client.FunctionalTests/HybridTimeTests.cs
@@ -84,7 +84,7 @@ namespace Knet.Kudu.Client.FunctionalTests
                 long snapshotTime = HybridTimeUtil.PhysicalAndLogicalToHtTimestamp(
                     futureTs, logicalValue);
                 int expected = i + 1;
-                int numRows = await ScanAtSnapshotAsync(snapshotTime);
+                long numRows = await ScanAtSnapshotAsync(snapshotTime);
                 Assert.Equal(expected, numRows);
             }
 
@@ -95,11 +95,11 @@ namespace Knet.Kudu.Client.FunctionalTests
             _client.LastPropagatedTimestamp++;
             long snapshotTime2 = HybridTimeUtil.PhysicalAndLogicalToHtTimestamp(
                 futureTs, latestLogicalValue + 1);
-            int numRows2 = await ScanAtSnapshotAsync(snapshotTime2);
+            long numRows2 = await ScanAtSnapshotAsync(snapshotTime2);
             Assert.Equal(1 + keys.Length, numRows2);
         }
 
-        private Task<int> ScanAtSnapshotAsync(long time)
+        private Task<long> ScanAtSnapshotAsync(long time)
         {
             var scanner = _client.NewScanBuilder(_table)
                 .SnapshotTimestampRaw(time)
